requirements.txt
# Function dependencies, for example:
# package>=version

line-bot-sdk
flask
gunicorn
